Refactor audio client into reusable core
