+++ Client/Source/AudioService.h
@@ -7,6 +7,7 @@
 #pragma once
 
 #include "AudioCallback.h"
+#include "JammerNetzSession.h"
 #include "ApplicationState.h"
 #include "DebounceTimer.h"
 
@@ -63,11 +64,16 @@ private:
 	std::shared_ptr<ChannelSetup> getSetup(ValueTree data) const;
 
 	void refreshChannelSetup(std::shared_ptr<ChannelSetup> setup);
+	void refreshEngineConfiguration();
+	void refreshSessionConfiguration();
+	JammerNetzSessionConfiguration getSessionConfiguration() const;
 	void restartAudio();
 	void restartAudio(std::shared_ptr<ChannelSetup> inputSetup, std::shared_ptr<ChannelSetup> outputSetup);
 
 	std::shared_ptr<juce::AudioIODevice> audioDevice_;
 
+	JammerNetzSession session_;
+	JammerNetzAudioEngine engine_;
 	AudioCallback callback_;
 	DebounceTimer debouncer_;
 	std::atomic<bool> shutdown_ { false };
