+++ Client/Source/DataReceiveThread.cpp
@@ -8,35 +8,11 @@
 
 #include "StreamLogger.h"
 
-#include "Data.h"
-#include "Encryption.h"
 #include "XPlatformUtils.h"
 
 DataReceiveThread::DataReceiveThread(DatagramSocket &socket, std::function<void(std::shared_ptr<JammerNetzAudioData>)> newDataHandler)
 	: Thread("ReceiveDataFromServer"), socket_(socket), newDataHandler_(newDataHandler), currentRTT_(0.0), isReceiving_(false), receiveErrorCount_(0), currentSession_(false)
 {
-	// Create listeners to get notified if the application state we depend on changes
-	listeners_.push_back(std::make_unique<ValueListener>(Data::getPropertyAsValue(VALUE_CRYPTOPATH), [this](Value& value) {
-		std::shared_ptr<MemoryBlock> cryptokey;
-		String newCryptopath = value.getValue();
-		if (newCryptopath.isNotEmpty()) {
-			UDPEncryption::loadKeyfile(newCryptopath.toRawUTF8(), &cryptokey);
-			if (cryptokey) {
-				setCryptoKey(cryptokey->getData(), safe_sizet_to_int(cryptokey->getSize()));
-			}
-			else {
-				// Turn off encrpytion
-				setCryptoKey(nullptr, 0);
-			}
-		}
-		else {
-			// Turn off encrpytion
-			setCryptoKey(nullptr, 0);
-		}
-	}));
-
-	// To read the current state, just execute each of these listeners once!
-	std::for_each(listeners_.begin(), listeners_.end(), [](std::unique_ptr<ValueListener>& ptr) { ptr->triggerOnChanged();  });
 }
 
 DataReceiveThread::~DataReceiveThread()
